backend/requirements.txt
Flask
google-cloud-vision
opencv-python